Add Yahoo! Weather forecast provider.
Adds a forecast provider that uses Yahoo! Weather RSS channels to
retrieve informations. For the time being, this requires a new tag
in the database, <yahoo-woeid>, containing a Where on Earth ID (the
one for Milan is provided for testing purposes).
The matchings from Yahoo condition codes to GWeatherConditions and
GWeatherSky need to be checked.