Stop using one XML file per language
The XML files are huge and take hours to build. Also, once they're
built we lose the information on the untranslated name.
Instead, use one XML file with the C locale names and regular
gettext to translate.
There is a bit of build system plumbing involved, because we want
to keep using a separate domain and po/ subdirectory (so that no
translations are lost)