locations: Correct Mexico's airport name
The airport for Mexico City is listed as "Licenci Airport", which is
incorrect.  The correct name of the airport is either "Benito Juárez
International Airport" or "Mexico City International Airport".  The
incorrect "Licenci Airport" may have come about as a truncation of
"Licenciado Benito Juárez International Airport", where Juárez's name is
given with his title, "Licenciado".  Adding the title "Licenciado" is
sometimes done in the name of the airport, but it is extremely formal, and
leaving it out is more common.

So, "Licenci Airport" is very wrong; "Benito Juárez International Airport"
or "Mexico City International Airport" should be used instead.

With help from brett@brettlangston.com

https://bugzilla.gnome.org/show_bug.cgi?id=625270