Remove GTK3 widgets
The GTK3 widgets cannot exist in the same process as other versions of
GTK, which means applications cannot port to GTK4 and link to
libgweather.

If we want to add GTK widgets back to libgweather we are going to do so
in a separate shared library, with a separate namespace.