Add Åre Östersund Airport and Östersund, Sweden
Though Östersund has less than 100k people, it is the capital of
Jämtland county (and the only "major" city in the area). It's
submitted for inclusion due to its significance for the region, and Åre
being a large tourist destination in the country most notably for skiing
and mountainbiking.