Allow using gweather-locations as a subproject
The gweather-locations project provides the locations database through
a pkg-config variable, but that's only available after installation. In
order to make this work when using subprojects, we need to provide the
variable as an overridden dependency.

Since the libgweather project uses the locations database in its tests
through an override, we also need to provide the path to the built
locations database, so that it's not necessary to install
gweather-locations when testing it.